requirements.development.txt
-r requirements.common.txt
ipython==7.31.0
coverage==6.2
factory-boy==3.2.1
selenium==4.1.0
django-nose==1.4.7
black==21.12b0
isort==5.10.1
